requirements.txt
blinker==1.7.0
certifi==2023.7.22
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.10.0
defusedxml==0.7.1
Flask==3.0.3
fonttools==4.51.0
gunicorn==22.0.0
idna==2.10
itsdangerous==2.2.0
Jinja2==3.1.3
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.4
numpy==1.26.4
opencv-python==4.9.0.80
opencv-python-headless==4.8.0.74
packaging==24.0
pandas==2.2.2
pillow==10.3.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-magic==0.4.27
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
requests-toolbelt==1.0.0
roboflow==1.1.27
scipy==1.13.0
six==1.16.0
supervision==0.19.0
tqdm==4.66.2
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.2
